# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Numerology CMS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RuleTypeCount {
  type: string;
  count: number;
}

export type RuleConditionJson = { [key: string]: unknown };

export type RuleResultJson = { [key: string]: unknown };

export interface Rule {
  id: number;
  name: string;
  rule_type: string;
  /** @nullable */
  description?: string | null;
  condition_json: RuleConditionJson;
  result_json: RuleResultJson;
  priority: number;
  is_active: boolean;
  created_at: string;
  updated_at: string;
}

export interface DashboardStats {
  totalRules: number;
  activeRules: number;
  totalFormulas: number;
  activeFormulas: number;
  totalKnowledgeEntries: number;
  totalLoShuRules: number;
  rulesByType: RuleTypeCount[];
  recentRules: Rule[];
}

export type RuleInputConditionJson = { [key: string]: unknown };

export type RuleInputResultJson = { [key: string]: unknown };

export interface RuleInput {
  name: string;
  rule_type: string;
  description?: string;
  condition_json: RuleInputConditionJson;
  result_json: RuleInputResultJson;
  priority?: number;
  is_active?: boolean;
}

export type RuleUpdateConditionJson = { [key: string]: unknown };

export type RuleUpdateResultJson = { [key: string]: unknown };

export interface RuleUpdate {
  name?: string;
  rule_type?: string;
  description?: string;
  condition_json?: RuleUpdateConditionJson;
  result_json?: RuleUpdateResultJson;
  priority?: number;
  is_active?: boolean;
}

export interface Formula {
  id: number;
  name: string;
  formula_type: string;
  formula_expression: string;
  /** @nullable */
  description?: string | null;
  is_active: boolean;
  created_at: string;
  updated_at: string;
}

export interface FormulaInput {
  name: string;
  formula_type: string;
  formula_expression: string;
  description?: string;
  is_active?: boolean;
}

export interface FormulaUpdate {
  name?: string;
  formula_type?: string;
  formula_expression?: string;
  description?: string;
  is_active?: boolean;
}

export interface KnowledgeEntry {
  id: number;
  title: string;
  category: string;
  content: string;
  tags: string[];
  is_published: boolean;
  created_at: string;
  updated_at: string;
}

export interface KnowledgeEntryInput {
  title: string;
  category: string;
  content: string;
  tags?: string[];
  is_published?: boolean;
}

export interface KnowledgeEntryUpdate {
  title?: string;
  category?: string;
  content?: string;
  tags?: string[];
  is_published?: boolean;
}

export interface MissingNumberRule {
  id: number;
  missing_number: number;
  title: string;
  strengths: string[];
  weaknesses: string[];
  recommendations: string[];
  is_active: boolean;
  created_at: string;
  updated_at: string;
}

export interface MissingNumberRuleInput {
  missing_number: number;
  title: string;
  strengths: string[];
  weaknesses: string[];
  recommendations: string[];
  is_active?: boolean;
}

export interface MissingNumberRuleUpdate {
  missing_number?: number;
  title?: string;
  strengths?: string[];
  weaknesses?: string[];
  recommendations?: string[];
  is_active?: boolean;
}

export interface RepeatedNumberRule {
  id: number;
  number: number;
  count: number;
  title: string;
  strengths: string[];
  weaknesses: string[];
  is_active: boolean;
  created_at: string;
  updated_at: string;
}

export interface RepeatedNumberRuleInput {
  number: number;
  count: number;
  title: string;
  strengths: string[];
  weaknesses: string[];
  is_active?: boolean;
}

export interface RepeatedNumberRuleUpdate {
  number?: number;
  count?: number;
  title?: string;
  strengths?: string[];
  weaknesses?: string[];
  is_active?: boolean;
}

export interface ArrowRule {
  id: number;
  name: string;
  numbers: number[];
  interpretation: string;
  is_active: boolean;
  created_at: string;
  updated_at: string;
}

export interface ArrowRuleInput {
  name: string;
  numbers: number[];
  interpretation: string;
  is_active?: boolean;
}

export interface ArrowRuleUpdate {
  name?: string;
  numbers?: number[];
  interpretation?: string;
  is_active?: boolean;
}

export interface ReportRequest {
  name: string;
  /** ISO date string: YYYY-MM-DD */
  date_of_birth: string;
}

export type NumerologyReportSubject = {
  name: string;
  date_of_birth: string;
};

export interface CalculatedNumbers {
  birthday_number: number;
  destiny_number: number;
  /** @nullable */
  personality_number?: number | null;
  personal_year: number;
  personal_month: number;
  personal_day: number;
}

export interface RuleInterpretation {
  rule_id: number;
  rule_name: string;
  rule_type: string;
  keywords: string[];
  strengths: string[];
  weaknesses: string[];
  recommendations: string[];
  /** @nullable */
  description?: string | null;
}

export interface RepeatedEntry {
  number: number;
  count: number;
}

export interface LoShuResult {
  grid: number[][];
  missing_numbers: number[];
  repeated_numbers: RepeatedEntry[];
  active_arrows: string[];
  missing_interpretations: MissingNumberRule[];
  repeated_interpretations: RepeatedNumberRule[];
  arrow_interpretations: ArrowRule[];
}

export interface NumerologyReport {
  subject: NumerologyReportSubject;
  numbers: CalculatedNumbers;
  interpretations: RuleInterpretation[];
  lo_shu: LoShuResult;
  generated_at: string;
}

export type ListRulesParams = {
rule_type?: string;
is_active?: boolean;
search?: string;
};

export type ListFormulasParams = {
is_active?: boolean;
};

export type ListKnowledgeEntriesParams = {
category?: string;
search?: string;
};