# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Numerology CMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /formulas:
    get:
      operationId: listFormulas
      tags: [formulas]
      summary: List all formulas
      parameters:
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Formulas list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Formula"
    post:
      operationId: createFormula
      tags: [formulas]
      summary: Create a formula
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormulaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Formula"
  /formulas/{id}:
    get:
      operationId: getFormula
      tags: [formulas]
      summary: Get a formula by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Formula
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Formula"
    patch:
      operationId: updateFormula
      tags: [formulas]
      summary: Update a formula
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormulaUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Formula"
    delete:
      operationId: deleteFormula
      tags: [formulas]
      summary: Delete a formula
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /formulas/{id}/duplicate:
    post:
      operationId: duplicateFormula
      tags: [formulas]
      summary: Duplicate a formula
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Formula"
  /formulas/{id}/toggle:
    patch:
      operationId: toggleFormula
      tags: [formulas]
      summary: Toggle formula active status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Formula"

  # ─── Knowledge Base ──────────────────────────────────────────────────────────
  /knowledge-entries:
    get:
      operationId: listKnowledgeEntries
      tags: [knowledge]
      summary: List knowledge base entries
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Entries list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeEntry"
    post:
      operationId: createKnowledgeEntry
      tags: [knowledge]
      summary: Create a knowledge entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeEntry"
  /knowledge-entries/{id}:
    get:
      operationId: getKnowledgeEntry
      tags: [knowledge]
      summary: Get a knowledge entry by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeEntry"
    patch:
      operationId: updateKnowledgeEntry
      tags: [knowledge]
      summary: Update a knowledge entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeEntryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeEntry"
    delete:
      operationId: deleteKnowledgeEntry
      tags: [knowledge]
      summary: Delete a knowledge entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Lo Shu ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Formula:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        formula_type:
          type: string
        formula_expression:
          type: string
        description:
          type: ["string", "null"]
        is_active:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - name
        - formula_type
        - formula_expression
        - is_active
        - created_at
        - updated_at
    FormulaInput:
      type: object
      properties:
        name:
          type: string
        formula_type:
          type: string
        formula_expression:
          type: string
        description:
          type: string
        is_active:
          type: boolean
      required:
        - name
        - formula_type
        - formula_expression
    FormulaUpdate:
      type: object
      properties:
        name:
          type: string
        formula_type:
          type: string
        formula_expression:
          type: string
        description:
          type: string
        is_active:
          type: boolean
    KnowledgeEntry:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        is_published:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - title
        - category
        - content
        - tags
        - is_published
        - created_at
        - updated_at
    KnowledgeEntryInput:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        is_published:
          type: boolean
      required:
        - title
        - category
        - content
    KnowledgeEntryUpdate:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        is_published:
          type: boolean
...