# lib/api-zod/src/generated/types/formulaInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Numerology CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface FormulaInput {
  name: string;
  formula_type: string;
  formula_expression: string;
  description?: string;
  is_active?: boolean;
}