# lib/api-zod/src/generated/types/knowledgeEntryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Numerology CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface KnowledgeEntryInput {
  title: string;
  category: string;
  content: string;
  tags?: string[];
  is_published?: boolean;
}